+++ utils/install/index.ts
@@ -1,9 +1,11 @@
+export { ALLOW_REMOTE_INSTALL } from './allowRemoteInstallFlag'
 export { defaultAccount } from './defaultAccount'
 export { DEFAULT_DAYS_BACK } from './defaultDaysBack'
 export { DEFAULT_IMAP_PORT } from './defaultImapPort'
 export { DEFAULT_INTERVAL } from './defaultInterval'
 export { installReducer } from './installReducer'
 export { MIN_SECRET_LENGTH } from './minSecretLength'
+export { timingSafeStringEqual } from './timingSafeStringEqual'
 
 export * from '../../validators/install/validateInstallBody'
 export * from './generateSecretKey'